requirements.txt
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
dotenv==0.9.9
fastapi==0.116.1
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.26.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.2
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==15.0.1
pydantic_settings==2.10.1
fastapi
uvicorn
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
requests
streamlit
